# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Dokümantasyonu",
        "description": "Bu API, sipariş oluşturma ve yönetme işlemleri için kullanılır.",
        "contact": {
            "email": "support@example.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:9002",
            "description": "Local Geliştirme Sunucusu"
        }
    ],
    "paths": {
...
        "/api/customer": {
            "get": {
                "tags": [
                    "Customer"
                ],
                "summary": "Get all customers",
                "description": "Fetches a list of all customers.",
                "operationId": "getCustomerAll",
                "responses": {
                    "200": {
                        "description": "List of customers",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "type": "integer",
                                                "example": 1
                                            },
                                            "name": {
                                                "type": "string",
                                                "example": "Irmak"
                                            },
                                            "since": {
                                                "type": "string",
                                                "format": "date",
                                                "example": "2025-01-01"
                                            },
                                            "revenue": {
                                                "type": "number",
                                                "format": "float",
                                                "example": 250.5
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Customers not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Müşteriler bulunamadı."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Customer"
                ],
                "summary": "Create an customer",
                "description": "Create a new customer and return the created customer.",
                "operationId": "createCustomer",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "since",
                                    "revenue"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Irmak"
                                    },
                                    "since": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2025-01-01"
                                    },
                                    "revenue": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 250.5
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Customer created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Müşteri başarıyla oluşturulmuştur."
                                        },
                                        "customer_id": {
                                            "type": "integer",
                                            "example": 123
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid input",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Invalid request data."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/customer/{id}": {
            "delete": {
                "tags": [
                    "Customer"
                ],
                "summary": "Delete an customer",
                "description": "Deletes an existing customer by ID.",
                "operationId": "deleteCustomer",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Customer ID",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 123
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Customer deleted successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Müşteri silindi"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Customer not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Müşteri Bulunamadı"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/discounted/{orderId}": {
            "get": {
                "tags": [
                    "Discounted"
                ],
                "summary": "Get order discounts",
                "description": "Fetch order discounts for the given order ID.",
                "operationId": "orderDiscounted",
                "parameters": [
                    {
                        "name": "orderId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 36
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Order discounts fetched successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": 36
                                        },
                                        "discounts": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "discountReason": {
                                                        "type": "string",
                                                        "example": "10_PERCENT_OVER_1000"
                                                    },
                                                    "discountAmount": {
                                                        "type": "string",
                                                        "example": "181.53"
                                                    },
                                                    "subtotal": {
                                                        "type": "string",
                                                        "example": "1,633.77"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        },
                                        "total": {
                                            "type": "string",
                                            "example": "1,815.30"
                                        },
                                        "totalDiscount": {
                                            "type": "string",
                                            "example": "291.81"
                                        },
                                        "discountedTotal": {
                                            "type": "string",
                                            "example": "1,523.49"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid order ID"
                    },
                    "422": {
                        "description": "Order not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}